Added an alternative function to check for parking spots
The added function uses trigonometric calculation to get the spot length, instead of moving the car. The function does not return a double, but instead returns a boolean.
The function also assumes that there is a sensor placed at the right side of the car, and also one at the right-front corner, placed diagonally.